move agents into a `custom` subfolder for folder structure consistency
