+++ src/modules/agents/custom/__init__.py
@@ -1,0 +1,2 @@
+from .custom_agent import CustomAgent
+from .ns_agent import NSAgent
